tz: break apart `src/tz/mod.rs`
This file was getting a bit big. And it was also responsible for
assembling a public API for `jiff::tz`. Mixing these two together just
made it a little hard for me to deal with.

I'm doing this with an eye toward #256, which will likely require
rejiggering the representation of `TimeZone` to use pointer tagging.
This will require `unsafe`, so I'd like to lock down its representation
to a smaller scope.
